# apispec.json
{
  "definitions": {
...
    "get_score_reply": {
      "properties": {
        "learning_name": {
          "type": "integer"
        }
      }
    }, 
...
    "leader_boards_reply": {
      "properties": {
        "comment_desc": {
          "type": "string"
        }, 
        "first_name": {
          "type": "string"
        }, 
        "last_name": {
          "type": "string"
        }, 
        "rating": {
          "type": "string"
        }, 
        "time_stamp": {
          "type": "object"
        }
      }
    }, 
    "learning_content_reply": {
      "properties": {
        "description": {
          "type": "string"
        }, 
        "learning_id": {
          "type": "string"
        }, 
        "link_to_video": {
          "type": "string"
        }, 
        "name": {
          "type": "string"
        }, 
        "skill_level": {
          "type": "string"
        }, 
        "tags": {
          "type": "string"
        }, 
        "topic": {
          "type": "string"
        }, 
        "topic_id": {
          "type": "string"
        }
      }
    }, 
    "learning_search_reply": {
      "properties": {
        "description": {
          "type": "string"
        }, 
        "id": {
          "type": "string"
        }, 
        "name": {
          "type": "string"
        }
      }
    }, 
...
    "upload_score_reply": {
      "properties": {
        "status": {
          "default": [
            "success", 
            "fail"
          ], 
          "type": "string"
        }
      }
    }, 
...
  }, 
  "info": {
    "description": "", 
    "title": "The Learning Hub API"
  }, 
  "paths": {
...
    "/get_score": {
      "get": {
        "description": "Call this api to get a score<br/>", 
        "parameters": [
          {
            "description": "The activity ID", 
            "in": "path", 
            "name": "activity_id", 
            "required": true, 
            "type": "integer"
          }
        ], 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/get_score_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the get score API", 
        "tags": [
          "API to get score"
        ]
      }
    }, 
...
    "/leader_boards": {
      "get": {
        "description": "Call this api to get leader boards<br/>", 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/leader_boards_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the leader boards API", 
        "tags": [
          "API to get leader boards"
        ]
      }
    }, 
    "/learning_API/{learning_id}": {
      "get": {
        "description": "Call this api passing a learning id and get back its content<br/>", 
        "parameters": [
          {
            "description": "The learning ID", 
            "in": "path", 
            "name": "learning_id", 
            "required": true, 
            "type": "integer"
          }
        ], 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/learning_content_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the get learning content API", 
        "tags": [
          "API to get learning content"
        ]
      }
    }, 
...
    "/search_learning_activity/{search_term}": {
      "get": {
        "description": "Call this api passing a search term and get back its result<br/>", 
        "parameters": [
          {
            "description": "The search term", 
            "in": "path", 
            "name": "search_term", 
            "required": true, 
            "type": "string"
          }
        ], 
        "responses": {
          "200": {
            "description": "Success", 
            "schema": {
              "$ref": "#/definitions/learning_search_reply"
            }
          }, 
          "500": {
            "description": "Error"
          }
        }, 
        "summary": "This is the search learning activity content API", 
        "tags": [
          "API to search learning activity"
        ]
      }
    }, 
...
    "/upload_score": {
      "post": {
        "description": "Call this api to upload score<br/>", 
        "parameters": [
          {
            "description": "The activity ID", 
            "in": "body", 
            "name": "activity_id", 
            "required": true, 
            "schema": {
              "properties": {
                "activity_id": {
                  "type": "integer"
                }, 
                "score": {
                  "type": "integer"
                }
              }
            }, 
            "type": "integer"
          }
        ], 
        "responses": {
          "200": {
            "description": "success", 
            "schema": {
              "$ref": "#/definitions/upload_score_reply"
            }
          }, 
          "400": {
            "description": "fail"
          }
        }, 
        "summary": "This is the upload score API", 
        "tags": [
          "API to upload score"
        ]
      }
    }, 
...
  }, 
  "swagger": "2.0"
}